feat: display a tab cell in the collection view
This concludes "Chapter 3 - Displaying A Tab Cell in the Collection
View". This involved creating a basic Tab class, a bespoke TabCell view,
and altering the TabCollectionVC to use that view. For now, the frame of
each cell is being hard-coded, just so we can see what the cell looks
like, but we'll want to size/position it more intelligently.